Revert "Merge remote-tracking branch 'refs/remotes/origin/while-loop' into while-loop"
This reverts commit 26d745d6c592fa92038f99f1eda9bbfef089d16e.
