+++ src/main/kotlin/com/egern/Main.kt
@@ -8,14 +8,9 @@ import com.egern.ast.Program
 import com.egern.codegen.CodeGenerationVisitor
 import com.egern.codegen.PreCodeGenerationVisitor
 import com.egern.emit.Emitter
-import com.egern.emit.LinuxEmitter
-import com.egern.emit.MacOSEmitter
-import com.egern.emit.WindowsEmitter
 import com.egern.error.ErrorLogger
 import com.egern.symbols.SymbolVisitor
 import com.egern.types.TypeCheckingVisitor
-import com.egern.util.Platform
-import com.egern.util.PlatformManager
 import com.egern.visitor.PrintProgramVisitor
 import com.egern.visitor.PrintSymbolTableVisitor
 import org.antlr.v4.runtime.CharStreams
@@ -60,22 +55,17 @@ fun main(args: Array<String>) {
     val preCodeGenerationVisitor = PreCodeGenerationVisitor()
     ast.accept(preCodeGenerationVisitor)
 
-    val platform = PlatformManager()
-
     val codeGenVisitor = CodeGenerationVisitor(symbolVisitor.currentTable)
     ast.accept(codeGenVisitor)
 
-    val emitter: Emitter = when (platform.platform) {
-        Platform.Windows -> WindowsEmitter(codeGenVisitor.instructions)
-        Platform.MacOS -> MacOSEmitter(codeGenVisitor.instructions)
-        Platform.Linux -> LinuxEmitter(codeGenVisitor.instructions)
-    }
+    val emitter = Emitter(codeGenVisitor.instructions)
     val code = emitter.emit()
+
     if (ErrorLogger.hasErrors()) {
         ErrorLogger.print()
         throw Exception("One or more errors occurred while compiling")
     } else if (!doPrint) {
-        print(code)
+        print(code.toString())
     }
 }
